chore: move Page.close() tests to tests/library
Page.close is not supported in WebView2, MCP extension and Android
