+++ tests/page/page-add-locator-handler.spec.ts
@@ -121,23 +121,6 @@ test('should not work with force:true', async ({ page, server }) => {
   expect(await page.evaluate('window.clicked')).toBe(undefined);
 });
 
-test('should throw when page closes', async ({ page, server, isAndroid }) => {
-  test.fixme(isAndroid, 'GPU process crash: https://issues.chromium.org/issues/324909825');
-  await page.goto(server.PREFIX + '/input/handle-locator.html');
-
-  await page.addLocatorHandler(page.getByText('This interstitial covers the button'), async () => {
-    await page.close();
-  });
-
-  await page.locator('#aside').hover();
-  await page.evaluate(() => {
-    (window as any).clicked = 0;
-    (window as any).setupAnnoyingInterstitial('mouseover', 1);
-  });
-  const error = await page.locator('#target').click().catch(e => e);
-  expect(error.message).toContain(kTargetClosedErrorMessage);
-});
-
 test('should throw when handler times out', async ({ page, server }) => {
   await page.goto(server.PREFIX + '/input/handle-locator.html');
