+++ tests/page/page-basic.spec.ts
@@ -18,134 +18,6 @@
 import { kTargetClosedErrorMessage } from '../config/errors';
 import { test as it, expect } from './pageTest';
 
-it('should reject all promises when page is closed', async ({ page, isWebView2, isAndroid }) => {
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-  it.fixme(isAndroid, '"Target crashed" instead of "Target closed"');
-
-  let error = null;
-  await Promise.all([
-    page.evaluate(() => new Promise(r => {})).catch(e => error = e),
-    page.close(),
-  ]);
-  expect(error.message).toContain(kTargetClosedErrorMessage);
-});
-
-it('should set the page close state', async ({ page, isWebView2 }) => {
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-
-  expect(page.isClosed()).toBe(false);
-  await page.close();
-  expect(page.isClosed()).toBe(true);
-});
-
-it('should pass page to close event', async ({ page, isAndroid, isWebView2 }) => {
-  it.fixme(isAndroid);
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-
-  const [closedPage] = await Promise.all([
-    page.waitForEvent('close'),
-    page.close()
-  ]);
-  expect(closedPage).toBe(page);
-});
-
-it('should terminate network waiters', async ({ page, server, isAndroid, isWebView2 }) => {
-  it.fixme(isAndroid);
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-
-  const results = await Promise.all([
-    page.waitForRequest(server.EMPTY_PAGE).catch(e => e),
-    page.waitForResponse(server.EMPTY_PAGE).catch(e => e),
-    page.close()
-  ]);
-  for (let i = 0; i < 2; i++) {
-    const message = results[i].message;
-    expect(message).toContain(kTargetClosedErrorMessage);
-    expect(message).not.toContain('Timeout');
-  }
-});
-
-it('should be callable twice', async ({ page, isWebView2 }) => {
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-
-  await Promise.all([
-    page.close(),
-    page.close(),
-  ]);
-  await page.close();
-});
-
-it('should fire load when expected', async ({ page }) => {
-  await Promise.all([
-    page.goto('about:blank'),
-    page.waitForEvent('load'),
-  ]);
-});
-
-it('async stacks should work', async ({ page, server }) => {
-  server.setRoute('/empty.html', (req, res) => {
-    req.socket.end();
-  });
-  let error = null;
-  await page.goto(server.EMPTY_PAGE).catch(e => error = e);
-  expect(error).not.toBe(null);
-  expect(error.stack).toContain(__filename);
-});
-
-it('should provide access to the opener page', async ({ page }) => {
-  const [popup] = await Promise.all([
-    page.waitForEvent('popup'),
-    page.evaluate(() => window.open('about:blank')),
-  ]);
-  const opener = await popup.opener();
-  expect(opener).toBe(page);
-});
-
-it('should return null if parent page has been closed', async ({ page, isWebView2 }) => {
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-
-  const [popup] = await Promise.all([
-    page.waitForEvent('popup'),
-    page.evaluate(() => window.open('about:blank')),
-  ]);
-  await page.close();
-  const opener = await popup.opener();
-  expect(opener).toBe(null);
-});
-
-it('should fire domcontentloaded when expected', async ({ page }) => {
-  const navigatedPromise = page.goto('about:blank');
-  await page.waitForEvent('domcontentloaded');
-  await navigatedPromise;
-});
-
-it('should pass self as argument to domcontentloaded event', async ({ page }) => {
-  const [eventArg] = await Promise.all([
-    new Promise(f => page.on('domcontentloaded', f)),
-    page.goto('about:blank')
-  ]);
-  expect(eventArg).toBe(page);
-});
-
-it('should pass self as argument to load event', async ({ page }) => {
-  const [eventArg] = await Promise.all([
-    new Promise(f => page.on('load', f)),
-    page.goto('about:blank')
-  ]);
-  expect(eventArg).toBe(page);
-});
-
-it('should fail with error upon disconnect', async ({ page, isAndroid, isWebView2 }) => {
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-  it.fixme(isAndroid);
-
-  let error;
-  const waitForPromise = page.waitForEvent('download').catch(e => error = e);
-  await page.close();
-  await waitForPromise;
-  expect(error.message).toContain(kTargetClosedErrorMessage);
-});
-
 it('page.url should work', async ({ page, server }) => {
   expect(page.url()).toBe('about:blank');
   await page.goto(server.EMPTY_PAGE);
@@ -172,14 +44,6 @@ it('page.close should work with window.close', async function({ page }) {
   const newPage = await newPagePromise;
   const closedPromise = new Promise(x => newPage.on('close', x));
   await page.evaluate(() => window['newPage'].close());
-  await closedPromise;
-});
-
-it('page.close should work with page.close', async function({ page, isWebView2 }) {
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-
-  const closedPromise = new Promise(x => page.on('close', x));
-  await page.close();
   await closedPromise;
 });
