+++ tests/page/page-event-network.spec.ts
@@ -138,23 +138,3 @@ it('should resolve responses after a navigation', async ({ page, server, browser
   // the response should resolve to null, because the page navigated.
   expect(await responsePromise).toBe(null);
 });
-
-it('interrupt request.response() and request.allHeaders() on page.close', async ({ page, server, browserName }) => {
-  it.info().annotations.push({ type: 'issue', description: 'https://github.com/microsoft/playwright/issues/27227' });
-  server.setRoute('/one-style.css', (req, res) => {
-    res.setHeader('Content-Type', 'text/css');
-  });
-  const reqPromise = page.waitForRequest('**/one-style.css');
-  await page.goto(server.PREFIX + '/one-style.html', { waitUntil: 'domcontentloaded' });
-  const req = await reqPromise;
-  const respPromise = req.response().catch(e => e);
-  const headersPromise = req.allHeaders().catch(e => e);
-  await page.close();
-  expect((await respPromise).message).toContain(kTargetClosedErrorMessage);
-  // All headers are the same as "provisional" headers in Firefox.
-  if (browserName === 'firefox')
-    expect((await headersPromise)['user-agent']).toBeTruthy();
-  else
-    expect((await headersPromise).message).toContain(kTargetClosedErrorMessage);
-
-});
