+++ tests/page/page-event-popup.spec.ts
@@ -146,22 +146,6 @@ it('should work with clicking target=_blank and rel=noopener', async ({ page, se
   expect(await popup.evaluate(() => !!window.opener)).toBe(false);
 });
 
-it('should not treat navigations as new popups', async ({ page, server, isWebView2 }) => {
-  it.skip(isWebView2, 'Page.close() is not supported in WebView2');
-
-  await page.goto(server.EMPTY_PAGE);
-  await page.setContent('<a target=_blank rel=noopener href="/one-style.html">yo</a>');
-  const [popup] = await Promise.all([
-    page.waitForEvent('popup'),
-    page.click('a'),
-  ]);
-  let badSecondPopup = false;
-  page.on('popup', () => badSecondPopup = true);
-  await popup.goto(server.CROSS_PROCESS_PREFIX + '/empty.html');
-  await page.close();
-  expect(badSecondPopup).toBe(false);
-});
-
 it('should report popup opened from iframes', async ({ page, server, browserName }) => {
   await page.goto(server.PREFIX + '/frames/two-frames.html');
   const frame = page.frame('uno');
