+++ tests/page/page-network-response.spec.ts
@@ -109,26 +109,6 @@ it('should wait until response completes', async ({ page, server }) => {
   expect(await responseText).toBe('hello world!');
 });
 
-it('should reject response.finished if page closes', async ({ page, server }) => {
-  await page.goto(server.EMPTY_PAGE);
-  server.setRoute('/get', (req, res) => {
-    // In Firefox, |fetch| will be hanging until it receives |Content-Type| header
-    // from server.
-    res.setHeader('Content-Type', 'text/plain; charset=utf-8');
-    res.write('hello ');
-  });
-  // send request and wait for server response
-  const [pageResponse] = await Promise.all([
-    page.waitForEvent('response'),
-    page.evaluate(() => fetch('./get', { method: 'GET' })),
-  ]);
-
-  const finishPromise = pageResponse.finished().catch(e => e);
-  await page.close();
-  const error = await finishPromise;
-  expect(error.message).toContain('closed');
-});
-
 it('should return json', async ({ page, server }) => {
   const response = await page.goto(server.PREFIX + '/simple.json');
   expect(await response.json()).toEqual({ foo: 'bar' });
